% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midbrainType-package.R
\docType{package}
\name{midbrainType-package}
\alias{midbrainType}
\alias{midbrainType-package}
\title{midbrainType: Automated Cell Type Annotation for Midbrain Single-Cell and Spatial Data}
\description{
Assigns cell type labels to single-cell or spatial transcriptomic data from the midbrain, using a curated marker gene reference bundled inside the package. No external atlas or reference download is required. Intended to be extended to further brain regions over time; this release covers the midbrain only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Notty4/midbrainType}
  \item Report bugs at \url{https://github.com/Notty4/midbrainType/issues}
}

}
\author{
\strong{Maintainer}: Naz \email{you@example.com}

}
\keyword{internal}
